# backend/specs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for archery statistics tracking",
        "title": "Arch Stats API",
        "contact": {},
        "version": "v0"
    },
    "basePath": "/api/v0",
    "paths": {
...
        "/auth/google": {
            "post": {
                "description": "Authenticate via Google One Tap credential. For existing archers, sets the auth cookie and returns credentials. For new archers, returns registration requirement.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Google One Tap Login",
                "parameters": [
                    {
                        "description": "Google One Tap credential",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/GoogleOneTapRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/AuthAuthenticated"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "Authenticate via Google One Tap credential. For existing archers, sets the auth cookie and returns credentials. For new archers, returns registration requirement.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Google One Tap Login",
                "parameters": [
                    {
                        "description": "Google One Tap credential",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/GoogleOneTapRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/AuthAuthenticated"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes the auth cookie and revokes the active session token in the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "operationId": "logout_api_v0_auth_logout_post",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/LogoutResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the currently authenticated archer from the request context or token cookie",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Get Current User",
                "operationId": "get_current_user_api_v0_auth_me_get",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/AuthAuthenticated"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Register a new archer profile with Google One Tap credential and demographic details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Register",
                "operationId": "register_api_v0_auth_register_post",
                "parameters": [
                    {
                        "description": "Registration details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/AuthRegistrationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/AuthAuthenticated"
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/AuthAuthenticated"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ArcherRead": {
            "type": "object",
            "required": [
                "archer_id",
                "bowstyle",
                "created_at",
                "date_of_birth",
                "draw_weight",
                "email",
                "first_name",
                "gender",
                "google_subject",
                "last_login_at",
                "last_name"
            ],
            "properties": {
                "archer_id": {
                    "type": "string"
                },
                "bowstyle": {
                    "$ref": "#/definitions/Bowstyle"
                },
                "club_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "date_of_birth": {
                    "type": "string"
                },
                "draw_weight": {
                    "type": "number"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "gender": {
                    "$ref": "#/definitions/Gender"
                },
                "google_picture_url": {
                    "type": "string"
                },
                "google_subject": {
                    "type": "string"
                },
                "last_login_at": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                }
            }
        },
...
        "AuthAuthenticated": {
            "type": "object",
            "required": [
                "access_token",
                "archer",
                "expires_at",
                "status"
            ],
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "archer": {
                    "$ref": "#/definitions/ArcherRead"
                },
                "expires_at": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/AuthStatus"
                }
            }
        },
        "AuthRegistrationRequest": {
            "type": "object",
            "required": [
                "bowstyle",
                "credential",
                "date_of_birth",
                "draw_weight",
                "gender"
            ],
            "properties": {
                "bowstyle": {
                    "$ref": "#/definitions/Bowstyle"
                },
                "club_id": {
                    "type": "string"
                },
                "credential": {
                    "type": "string",
                    "minLength": 10
                },
                "date_of_birth": {
                    "type": "string"
                },
                "draw_weight": {
                    "type": "number",
                    "maximum": 200
                },
                "first_name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "gender": {
                    "$ref": "#/definitions/Gender"
                },
                "last_name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                }
            }
        },
        "AuthStatus": {
            "type": "string",
            "enum": [
                "authenticated",
                "needs_registration"
            ],
            "x-enum-varnames": [
                "AuthStatusAuthenticated",
                "AuthStatusNeedsRegistration"
            ]
        },
        "Bowstyle": {
            "type": "string",
            "enum": [
                "recurve",
                "compound",
                "barebow",
                "longbow"
            ],
            "x-enum-varnames": [
                "BowstyleRecurve",
                "BowstyleCompound",
                "BowstyleBarebow",
                "BowstyleLongbow"
            ]
        },
        "ErrorResponse": {
            "type": "object",
            "required": [
                "detail"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                }
            }
        },
...
        "Gender": {
            "type": "string",
            "enum": [
                "male",
                "female",
                "non_binary",
                "other",
                "unspecified"
            ],
            "x-enum-varnames": [
                "GenderMale",
                "GenderFemale",
                "GenderNonBinary",
                "GenderOther",
                "GenderUnspecified"
            ]
        },
        "GoogleOneTapRequest": {
            "type": "object",
            "required": [
                "credential"
            ],
            "properties": {
                "credential": {
                    "type": "string",
                    "minLength": 10
                }
            }
        },
...
        "LogoutResponse": {
            "type": "object",
            "required": [
                "success"
            ],
            "properties": {
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "access_token",
            "in": "cookie"
        }
    }
}